requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlitedict
faiss-cpu
sentence-transformers
openai
pandas
PyPDF2
python-docx
openpyxl
google-cloud-firestore
python-dotenv